no fixes

+++ include/airline.h
@@ -2,10 +2,14 @@
 #define _AIRLINE_
 
 #include "common.h"
+#include "plane.h"
+#include <sys/types.h>
 #include <string.h>
+#include <unistd.h>
+
 
 typedef struct {
-	long id;
+	int id;
 	int planeCount;
 	int* targetedCities;
 } Airline;
